perf: resmap: Add Transform function
Introduce a new function, Transform, which can be used to safely modify
entries an resmap.

The callback function is called for each entry in the map and handles ID
conflict detection, which was the most expensive operation causing
significant performance issues.

The ID cache is local to this function as we cannot guarantee that all
mutations are performed using this function.