perf: resmap: Add AppendMany function
Introduce a new function, AppendMany, which can be used to safely add
entries to a resmap.

The function handles ID conflict detection, which was the most expensive
operation causing significant performance issues.

The ID cache is local to this function as we cannot guarantee that all
mutations are performed using this function.